ini8_backend/requirements.txt
Flask
flask_sqlalchemy
flask_cors
psycopg2-binary
python-dotenv
PyJWT
pytest
redis
